+++ tests/test_webdriver_chrome.py
@@ -9,7 +9,7 @@ class ChromeBrowserTest(WebDriverTests, unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.browser = Browser('webdriver.chrome')
+        cls.browser = Browser('chrome')
 
     def setUp(self):
         self.browser.visit(EXAMPLE_APP)